requirements.txt
accelerate
diffusers==0.28.0
transformers==4.47.0
huggingface_hub==0.24.0
einops
decord
tqdm
safetensors
omegaconf
opencv-python
pydantic
compel
easydict
rotary_embedding_torch
imageio[ffmpeg]
gradio
httpx[socks]
numpy>1.2
hydra-core==1.1.1
hydra-colorlog
matplotlib
opencv-python
omegaconf
plotly
ftfy
torch==2.1.1
torchsde
torchdiffeq
vit-pytorch==1.4.4
pytorch-lightning==1.9.5
voltron-robotics
gitpython
scipy
sentence-transformers
torchsde
einops
torchdiffeq
gym
moviepy==1.0.3
tqdm
termcolor
wandb
einops_exts
lpips